+++ RELEASENOTES.md
@@ -17,6 +17,9 @@
 *   Track selection:
     *   Change signature of `DefaultTrackSelector.selectAllTracks` to include
         the output array as a parameter, pre-populated with track overrides.
+    *   Fix issue where embedded metadata tracks are selected irrespective of
+        whether the corresponding primary track is selected
+        ([#294](https://github.com/androidx/media/issues/294)).
 *   Extractors:
     *   MP4: Support files with large `uuid` boxes before other boxes like
         `moov` ([#3046](https://github.com/androidx/media/issues/3046)).